Update version.fi only needed
This commit updates the Python script to update version.fi only the
version information changed. This will prevent dependent projects from
being built unnecessarily.

+++ dsm2/src/common/version_generate.py
@@ -5,7 +5,6 @@ VersionTemplate     = "      character*16 :: dsm2_version = '@{Version_GIT_DESCR
 VersionFile_path    = os.path.split( __file__)[0]
 VersionFile_path    = os.path.join(VersionFile_path,"version.fi")
 
-VersionFile = open(VersionFile_path, "w")
 
 try:
     p = subprocess.Popen("git describe", shell=True, stdout=subprocess.PIPE, universal_newlines=True)
@@ -23,9 +22,23 @@ try:
     VersionTxt = VersionTemplate.replace("@{Version_GIT_DESCRIBE}", GITVersion_Describe)
     VersionTxt = VersionTxt.replace("@{Version_GIT}", GITVersion_SourceCode)
     VersionTxt = VersionTxt.replace("@{GUID_GIT}", GITGUID_SourceCode)
-    VersionFile.write(VersionTxt)
-    VersionFile.close()
+    if os.path.exists(VersionFile_path):
+        with open(VersionFile_path, "r") as f:
+            version_changed = False
+            VersionFile_old = f.readline()
+            if VersionFile_old == VersionTxt:
+                print('No version changes.')
+            else:
+                print('The version has changed.')
+                version_changed = True
+        if version_changed:
+            with open(VersionFile_path, "w") as f:
+                f.write(VersionTxt)
+    else:
+        print('Create a new version file.')
+        with open(VersionFile_path, "w") as f:
+            f.write(VersionTxt)
 except:
-    VersionFile.close()
-    os.remove(VersionFile_path)
+    if os.path.exists(VersionFile_path):
+        os.remove(VersionFile_path)
     print('Abort.... possible error in file /dsm2/src/common/version_generate.py')
